restAPI/requirements.txt
black==23.3.0
python-dotenv==1.0.0
requests==2.30.0
